pyproject.toml
[tool.poetry]
name = "homework-24-25"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
black = "^24.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
